Modified chat to hide the API key.  This is a combination of using the npm 'environment' package, a new route (because the environment package only works on the back end), and an ajax call in the chat.js routine to get the API information. It's not quite secure, but at least the API key isn't left out in the open.
