+++ server.js
@@ -26,12 +26,13 @@ var exphbs = require("express-handlebars");
 app.engine("handlebars", exphbs({ defaultLayout: "main" }));
 app.set("view engine", "handlebars");
 
+require("dotenv").config();
+
 // Routes
 // =============================================================
 require("./routes/html-routes.js")(app);
 require("./routes/match-api-routes.js")(app);
 require("./routes/user-api-routes.js")(app);
-// require("./routes/userinfo-api-routes.js")(app);
 
 // Syncing our sequelize models and then starting our Express app
 // =============================================================
